+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## Dodo x.x.x (not yet released)
 * Update to wikimedia/zest-css 2.0.0.
 * Update to wikimedia/idle-dom 0.8.0.
+* Update to wikimedia/remex-html 2.3.2.
 * Implement the following non-standard methods for PHP compatibility:
   * Document::loadHTML()
   * Document::loadXML()